requirements.txt
langchain
langchain-community
langchain-groq
langchain-core
pypdf
sentence-transformers
faiss-cpu
python-dotenv
streamlit